backend/pyproject.toml
[project]
name = "saveslate-backend"
version = "0.1.0"
description = "SaveSlate personal finance tracker – API backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.15,<2",
    "pyjwt>=2.10,<3",
    "bcrypt>=4.3,<5",
    "pydantic[email]>=2.11,<3",
    "pydantic-settings>=2.9,<3",
    "python-multipart>=0.0.20,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "httpx>=0.28,<1",
    "aiosqlite>=0.21,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
